% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_plot.R
\name{den_plot}
\alias{den_plot}
\title{create a density plot for the desired columns of the data set with a regression line using the empirical data}
\usage{
den_plot(sim_data, emp_col, sim_col, xlabel)
}
\arguments{
\item{sim_data}{is the simulated data set you want to use for the plot}

\item{emp_col}{is the column from the empirical data set}

\item{sim_col}{is the column you want to use from the data set (needs to be the same as the emp_col)}

\item{xlabel}{is the title of the X-Axis}
}
\value{
gives you the resulting plot.
}
\description{
create a density plot for the desired columns of the data set with a regression line using the empirical data
}
\examples{
plot <- den_plot(simdata, empdata$column1, simdata$column1, "Title of the X-Axis")
}
