% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_size.R
\name{effect_size}
\alias{effect_size}
\title{calculate the effect size of the column values}
\usage{
effect_size(sim_col, emp_col)
}
\arguments{
\item{sim_col}{is the column of the simulated data set}

\item{emp_col}{is the corresponding column of the empirical data set}
}
\value{
gives you the effect size of the column values
}
\description{
calculate the effect size of the column values
}
\examples{
effect_size(simdata$tajima_pi, empdata$tajima_pi)  
}
