% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model.R
\name{linear_model}
\alias{linear_model}
\title{Linear Model Testing between values of two columns}
\usage{
linear_model(path_to_data)
}
\arguments{
\item{path_to_data}{is the directory path where the data file is.}
}
\value{
It gives the summary of model_fit variable including R-Squared values and Intercepts.
}
\description{
Linear Model Testing between values of two columns
}
\examples{
linear_model("data/col_with_values")

}
