% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_blank.R
\name{remove_blank}
\alias{remove_blank}
\title{Remove the columns with no data in it}
\usage{
remove_blank(path_to_data)
}
\arguments{
\item{path_to_data}{is the directory path where the data file is.}
}
\value{
gives you the data frame containing columns that has data in it.
}
\description{
Remove the columns with no data in it
}
\examples{
remove_blank("data/cleaned_names_simulated.csv")

}
