% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_blank.R
\name{remove_blank}
\alias{remove_blank}
\title{Remove the columns with no data in it}
\usage{
remove_blank(data)
}
\arguments{
\item{data}{is the data set you want to remove the blank columns from}
}
\value{
gives you the data frame containing columns that has data in it.
}
\description{
Remove the columns with no data in it
}
\examples{
data_w_values <- remove_blank("mydata")
}
