% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_columns.R
\name{rename_col}
\alias{rename_col}
\title{Rename the column names for easy use}
\usage{
rename_col(path_to_data)
}
\arguments{
\item{path_to_data}{is the directory path where the data file is.}
}
\value{
gives you the data frame containing cleaned column names.
}
\description{
Rename the column names for easy use
}
\examples{
my_data <- rename_col("data/simulated_data_pps_data.csv")
}
